# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Opendata - OpenQuoteAPI 1.0
  description: |-
    Open Data - Open Quote server based on OpenAPI 3.0 specifications, following API First Principals. This project aims to provide endpoints for clients to send and receive insurance quote data from across the Open Data family.
  termsOfService: http://open.data/terms/
  version: "1.0"
  contact:
    email: pmaybin01@gmail.com
...
servers:
  - url: http://localhost:8080/
    description: Development server
  - url: https://aws.staging.opendata.io/app/v1
    description: Staging server
  - url: https://aws.prod.opendata.io/app/v1
    description: Production server
...
paths:
  /quote:
    put:
      tags:
        - quote
      summary: Update an existing quote
      description: Update an existing quote by Id
      operationId: updateQuote
      requestBody:
        description: Update an existent quote
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quote'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Quote'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
            application/xml:
              schema:
                $ref: '#/components/schemas/Quote'

        '400':
          description: Bad request - Invalid ID supplied
        '404':
          description: Not Found - Quote not found
        '405':
          description: Method not allowed - Validation exception
      security:
        - broker_auth:
            - write:quote
            - read:quote
    post:
      tags:
        - quote
      summary: Add a new pet to the store
      description: Add a new pet to the store
      operationId: addPet
      requestBody:
        description: Create a new pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quote'
          application/xml:
            schema:
              $ref: '#/components/schemas/Quote'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Quote'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
            application/xml:
              schema:
                $ref: '#/components/schemas/Quote'
        '405':
          description: Invalid input
      security:
        - petstore_auth:
            - write:quote
            - read:quote
...
components:
  schemas:
    Quote:
      required:
        - quote_number
        - street_address
        - city
        - zip_code
        - state
        - year_built
        - construction_type
        - date_created
        - quote_premium
      type: object
      properties:
        quote_number:
          type: integer
          format: int64
          example: 000001
        street_address:
          type: string
          example: 14064 Reindeer St
        city:
          type: string
          example: Springfield
        zip_code:
          type: integer
          format: int32
          example: 34666
        state:
          type: string
          example: FL
        year_built:
          type: integer
          format: int32
          example: 1978
        construction_type:
          type: string
          description: type of construction for building
          example:
            MASONARY
          enum:
            - MASONARY
            - FRAME
        date_created:
          type: string
          format: date-time
          example: 28/05/2022T15:37
        quote_premium:
          type: number
          format: double
          example: 1523.74
...
  securitySchemes:
    broker_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://aws.opendata.io/oauth/authorize
          scopes:
            write:quote: modify quote info
            read:quote: read quote information
    api_key:
      type: apiKey
      name: api_key
      in: header